requirements.txt
click==8.1.7
fastapi==0.115.12
gTTS==2.5.3
keyboard==0.13.5
langchain==0.3.24
langchain_chroma==0.2.3
langchain_community==0.3.22
langchain_core==0.3.56
langchain_ollama==0.3.2
langchain_text_splitters==0.3.8
llama_index==0.12.33
numpy==1.24.3
openai_whisper==20240930
pydantic==2.11.3
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.1.0
pyttsx3==2.98
scipy==1.15.2
silero_vad==5.1.2
sounddevice==0.4.6
sounddevice==0.5.1
SpeechRecognition==3.10.4
streamlit==1.38.0
streamlit_pdf_viewer==0.0.17
streamlit_webrtc==0.62.4
useful_moonshine_onnx==20241016
uvicorn==0.34.2
